se añade metodo findbyid en workboard y metodo eliminar, ruta del metodo findbyid

+++ backend/controllers/workBoard.js
@@ -1,4 +1,5 @@
 import workBoard from "../models/workBoard.js";
+import board from "../models/board.js";
 
 const saveWorkB = async(req, res) => {
     if (!req.body.name || !req.body.description)
@@ -38,11 +39,25 @@ const updateWorkB = async(req, res) => {
         res.status(200).send({ message: "Work update" });
 };
 
+
 const deleteWorkB = async(req, res) => {
+    const workId = await workBoard.findById({ _id: req.params["_id"] });
+    if (!workId) return res.status(400).send({ message: "Work board not found" });
+
+    const boardList = await board.deleteMany({ workBoardId: workId });
+
     const workDelete = await workBoard.findByIdAndDelete({ _id: req.params["_id"] });
     return !workDelete ?
-        res.status(400).send({ message: "user no found" }) :
+        res.status(400).send({ message: "Work board not found" }) :
         res.status(200).send({ message: "Work board deleted" });
 };
 
-export default { saveWorkB, listWorkB, updateWorkB, deleteWorkB };
+const findWork = async(req, res) => {
+    const workfind = await workBoard
+        .findById({ _id: req.params["_id"] });
+    return !workfind ?
+        res.status(400).send({ message: "No search results" }) :
+        res.status(200).send({ workfind });
+};
+
+export default { saveWorkB, listWorkB, updateWorkB, deleteWorkB, findWork };
